Refactor booted-device helpers in device module
Move booted-UDID lookup logic into `device.ts` as `listBootedUdids()` and export `eachDevice()` for shared iteration, then update CLI and middleware to use the centralized helpers. This removes duplicate fallback logic around device-set reads, keeps the null-on-read-failure behavior for destructive paths, and slightly simplifies simulator listing by spreading existing device fields.
